Fleshed out "Data" section.
+ Added the one sentence with DOI link per requirements on canvas.
+ Pasted in dataset definitions
Requirements from 'Project check-in 2: Identification of data, tools and infrastructure necessary':
"A section with the heading "Data". Under this heading briefly describe in at least one complete sentence where you have obtained the data for your replication. If your data are hosted on a repository, give the name of that repository. If your data have a DOI, provide it here as a link. If there's no repository and no DOI, include whatever link you can find that takes us as close as possible to the original data files."